# server/api/jsonrpc/v1/jsonrpc_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.0
// - protoc             v5.29.3
// source: api/jsonrpc/v1/jsonrpc.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationJsonrpcGetJsonrpc = "/jsonrpc.v1.Jsonrpc/GetJsonrpc"
const OperationJsonrpcPostJsonrpc = "/jsonrpc.v1.Jsonrpc/PostJsonrpc"

type JsonrpcHTTPServer interface {
	// GetJsonrpc GET /rpc/{url}?id=xxx&jsonrpc=2.0&method=ping 之类
	GetJsonrpc(context.Context, *GetJsonrpcRequest) (*GetJsonrpcReply, error)
	// PostJsonrpc POST /rpc/{url}
	// body 是 JSON-RPC 报文的“结构化版本”
	PostJsonrpc(context.Context, *PostJsonrpcRequest) (*PostJsonrpcReply, error)
}

func RegisterJsonrpcHTTPServer(s *http.Server, srv JsonrpcHTTPServer) {
	r := s.Route("/")
	r.GET("/rpc/{url}", _Jsonrpc_GetJsonrpc0_HTTP_Handler(srv))
	r.POST("/rpc/{url}", _Jsonrpc_PostJsonrpc0_HTTP_Handler(srv))
}

func _Jsonrpc_GetJsonrpc0_HTTP_Handler(srv JsonrpcHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetJsonrpcRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationJsonrpcGetJsonrpc)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetJsonrpc(ctx, req.(*GetJsonrpcRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetJsonrpcReply)
		return ctx.Result(200, reply)
	}
}

func _Jsonrpc_PostJsonrpc0_HTTP_Handler(srv JsonrpcHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in PostJsonrpcRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationJsonrpcPostJsonrpc)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.PostJsonrpc(ctx, req.(*PostJsonrpcRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*PostJsonrpcReply)
		return ctx.Result(200, reply)
	}
}

type JsonrpcHTTPClient interface {
	// GetJsonrpc GET /rpc/{url}?id=xxx&jsonrpc=2.0&method=ping 之类
	GetJsonrpc(ctx context.Context, req *GetJsonrpcRequest, opts ...http.CallOption) (rsp *GetJsonrpcReply, err error)
	// PostJsonrpc POST /rpc/{url}
	// body 是 JSON-RPC 报文的“结构化版本”
	PostJsonrpc(ctx context.Context, req *PostJsonrpcRequest, opts ...http.CallOption) (rsp *PostJsonrpcReply, err error)
}

type JsonrpcHTTPClientImpl struct {
	cc *http.Client
}

func NewJsonrpcHTTPClient(client *http.Client) JsonrpcHTTPClient {
	return &JsonrpcHTTPClientImpl{client}
}

// GetJsonrpc GET /rpc/{url}?id=xxx&jsonrpc=2.0&method=ping 之类
func (c *JsonrpcHTTPClientImpl) GetJsonrpc(ctx context.Context, in *GetJsonrpcRequest, opts ...http.CallOption) (*GetJsonrpcReply, error) {
	var out GetJsonrpcReply
	pattern := "/rpc/{url}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationJsonrpcGetJsonrpc))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// PostJsonrpc POST /rpc/{url}
// body 是 JSON-RPC 报文的“结构化版本”
func (c *JsonrpcHTTPClientImpl) PostJsonrpc(ctx context.Context, in *PostJsonrpcRequest, opts ...http.CallOption) (*PostJsonrpcReply, error) {
	var out PostJsonrpcReply
	pattern := "/rpc/{url}"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationJsonrpcPostJsonrpc))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}